Create files app and migrate models and admins
